Move EdgeVerifierHelper from routing.rs to edge_verifier_actor.rs
